refactor(web): extract shared comments FE kit (S3a)
Move the domain-neutral half of the multitable comment UI into
apps/web/src/shared/comments/: MetaCommentsPanel, MetaCommentComposer,
MetaCommentReactions, meta-comment-labels.ts, the normalizeMultitableComment
wire-payload normalizer family, and the useMultitableComments composable —
byte-identical templates/styles/classes/testids/i18n strings, no redesign.

Introduces the new CommentsApiClient interface (the 7 methods the composable
actually calls), extracted from usage; the composable now depends on that
interface with a required client param instead of defaulting to the
multitableClient singleton. multitable/api/client.ts's MultitableApiClient
is the first implementation (`implements CommentsApiClient`); the
multitable-side useMultitableComments becomes a thin adapter supplying the
multitableClient default, preserving its external signature.

multitable/types.ts, multitable/api/client.ts, and
multitable/utils/meta-comment-labels.ts re-export the moved symbols under
their original names so every existing multitable consumer (views,
composables, specs) keeps resolving unchanged. multitable/components/
MetaComments{Panel,Composer}.vue and MetaCommentReactions.vue become
re-export shims at the old path, kept only because a handful of frozen specs
import/introspect them from there; living callers (MetaCommentsDrawer.vue,
MetaRecordInspector.vue) now import the shared kit directly.

MetaCommentActionChip.vue, MetaCommentAffordance.vue, and
utils/comment-affordance.ts are deliberately LEFT IN PLACE — they are woven
into the RATIFIED comment-affordance-color-consistency.spec.ts design-lock
guard, which hardcodes their old path and source-scans their own CSS/import
text alongside 8 other permanently-multitable view files. None of the three
moved components import comment-affordance.ts, so moving it would have
served no shared-kit consumer while breaking a frozen guard.

Adds tests/shared-comments-stub-client.spec.ts, proving the composable works
against a minimal hand-written CommentsApiClient stub with no multitable
dependency — the seam a future approval-native storage backend implements.
Wired into run-required-web-tests.sh (bare token, verified unique, 7/7
green).

No approval-side consumer, no authz, no backend change — kit extraction
only.
