+++ apps/web/src/multitable/api/client.ts
@@ -82,6 +82,31 @@ import type {
 import { apiFetch } from '../../utils/api'
 import { apiDefaultErrorMessage, apiFieldValidationFallback } from '../utils/meta-api-error-labels'
 import type { ConditionalRuleDTO } from '../utils/conditional-rule-ops'
+// S3a (comments shared FE kit extraction): the comment normalizer family moved verbatim to
+// shared/comments/normalize.ts. Imported here (for this file's own internal use in
+// normalizeCommentsList/normalizeCommentInbox below) AND re-exported under the same names, so
+// external consumers that import them from `'../api/client'` (comments-realtime.ts,
+// multitable-client.spec.ts) keep working unchanged. CommentsApiClient is the new interface
+// MultitableApiClient implements (below) as its first implementation.
+import type { RawComment } from '../../shared/comments/normalize'
+import {
+  normalizeMultitableComment,
+  normalizeMultitableCommentIdentity,
+  normalizeMultitableCommentFieldId,
+  normalizeMultitableCommentMentions,
+  normalizeMultitableCommentReactions,
+} from '../../shared/comments/normalize'
+import type { CommentsApiClient } from '../../shared/comments/api-client'
+
+// RawComment was not exported pre-move either — kept module-private, only re-exporting the
+// function family below (which WAS the public surface).
+export {
+  normalizeMultitableComment,
+  normalizeMultitableCommentIdentity,
+  normalizeMultitableCommentFieldId,
+  normalizeMultitableCommentMentions,
+  normalizeMultitableCommentReactions,
+}
 
 type FetchFn = (input: string, init?: RequestInit) => Promise<Response>
 type ApiErrorLocaleResolver = () => boolean
@@ -340,11 +365,6 @@ function parseContentDispositionFilename(header: string | null): string {
   return 'export'
 }
 
-type RawComment = Partial<MultitableComment> & {
-  spreadsheetId?: string
-  rowId?: string
-}
-
 type RawInboxItem = RawComment & {
   unread?: boolean
   mentioned?: boolean
@@ -359,108 +379,11 @@ type RawInboxItem = RawComment & {
   fieldName?: string | null
 }
 
-type MultitableCommentIdentityPayload = {
-  containerId?: unknown
-  spreadsheetId?: unknown
-  targetId?: unknown
-  rowId?: unknown
-}
-
-type MultitableCommentFieldPayload = {
-  fieldId?: unknown
-  targetFieldId?: unknown
-}
-
-type MultitableCommentMentionsPayload = {
-  mentions?: unknown
-}
-
-function normalizeCommentId(value: unknown): string {
-  if (typeof value !== 'string') return ''
-  const trimmed = value.trim()
-  return trimmed.length > 0 ? trimmed : ''
-}
-
-function normalizeOptionalCommentId(value: unknown): string | undefined {
-  const normalized = normalizeCommentId(value)
-  return normalized.length > 0 ? normalized : undefined
-}
-
-export function normalizeMultitableCommentIdentity(payload: MultitableCommentIdentityPayload | null | undefined) {
-  const containerId = normalizeCommentId(payload?.containerId) || normalizeCommentId(payload?.spreadsheetId)
-  const targetId = normalizeCommentId(payload?.targetId) || normalizeCommentId(payload?.rowId)
-
-  return {
-    containerId,
-    targetId,
-    spreadsheetId: containerId || undefined,
-    rowId: targetId || undefined,
-  }
-}
-
-export function normalizeMultitableCommentFieldId(payload: MultitableCommentFieldPayload | null | undefined): string | null {
-  const fieldId = normalizeCommentId(payload?.fieldId) || normalizeCommentId(payload?.targetFieldId)
-  return fieldId.length > 0 ? fieldId : null
-}
-
-export function normalizeMultitableCommentMentions(payload: MultitableCommentMentionsPayload | null | undefined): string[] {
-  if (!Array.isArray(payload?.mentions)) return []
-  return payload.mentions.filter((value): value is string => typeof value === 'string' && value.trim().length > 0)
-}
-
-/**
- * Normalize the per-comment `reactions` aggregate (B6). The whitelist normalizer
- * drops unknown raw fields, so reactions MUST be carried explicitly here or the
- * backend's reactions array is silently lost on the wire (wire-vs-fixture drift).
- * Returns undefined when absent (so a comment whose reactions weren't hydrated is
- * distinguishable from one with zero reactions).
- */
-export function normalizeMultitableCommentReactions(
-  payload: { reactions?: unknown } | null | undefined,
-): MultitableCommentReaction[] | undefined {
-  if (!Array.isArray(payload?.reactions)) return undefined
-  const out: MultitableCommentReaction[] = []
-  for (const raw of payload.reactions) {
-    if (!raw || typeof raw !== 'object') continue
-    const r = raw as Record<string, unknown>
-    if (typeof r.emoji !== 'string' || !r.emoji) continue
-    out.push({
-      emoji: r.emoji,
-      count: typeof r.count === 'number' && Number.isFinite(r.count) ? r.count : 0,
-      reactedByMe: r.reactedByMe === true,
-    })
-  }
-  return out
-}
-
 function normalizeCommentPresenceList(payload: { items?: MultitableCommentPresenceSummary[] } | null | undefined): {
   items: MultitableCommentPresenceSummary[]
 } {
   if (!payload?.items) return { items: [] }
   return { items: payload.items.map(normalizeMultitableCommentPresenceSummary) }
-}
-
-export function normalizeMultitableComment(payload: RawComment | null | undefined): MultitableComment {
-  const identity = normalizeMultitableCommentIdentity(payload)
-  const fieldId = normalizeMultitableCommentFieldId(payload)
-  return {
-    id: normalizeCommentId(payload?.id),
-    containerId: identity.containerId,
-    targetId: identity.targetId,
-    spreadsheetId: identity.spreadsheetId,
-    rowId: identity.rowId,
-    fieldId,
-    targetFieldId: fieldId,
-    parentId: normalizeOptionalCommentId(payload?.parentId),
-    mentions: normalizeMultitableCommentMentions(payload),
-    authorId: normalizeCommentId(payload?.authorId),
-    authorName: typeof payload?.authorName === 'string' ? payload.authorName : undefined,
-    content: typeof payload?.content === 'string' ? payload.content : '',
-    resolved: payload?.resolved === true,
-    createdAt: typeof payload?.createdAt === 'string' ? payload.createdAt : '',
-    updatedAt: typeof payload?.updatedAt === 'string' ? payload.updatedAt : undefined,
-    reactions: normalizeMultitableCommentReactions(payload),
-  }
 }
 
 function normalizeCommentsList(payload: { comments?: RawComment[]; items?: RawComment[] } | null | undefined): {
@@ -1427,7 +1350,11 @@ function configRestoreConfirmForPreview(preview: ConfigRestorePreview): ConfigRe
   return undefined
 }
 
-export class MultitableApiClient {
+// S3a: MultitableApiClient is the first implementation of the shared CommentsApiClient interface
+// (listComments/createComment/resolveComment/updateComment/deleteComment/addReaction/
+// removeReaction below already have this exact shape — `implements` only adds a compile-time
+// check, no runtime change).
+export class MultitableApiClient implements CommentsApiClient {
   private fetch: FetchFn
   private readonly isZhOption?: ApiErrorLocaleOption
   private readonly configRestoreConfirmByToken = new Map<string, ConfigRestoreExecuteConfirm>()
